More code cleanup.

+++ g_ctf.c
@@ -2351,7 +2351,7 @@ void CTFScoreboardMessage (edict_t *ent, edict_t *killer)
 {
 	char	entry[1024];
 	char	string[1400];
-	int		len;
+	size_t		len;
 	int		i, j, k, n;
 	//int		captures[2][MAX_CLIENTS];
 	int		sorted[2][MAX_CLIENTS];
@@ -2361,7 +2361,7 @@ void CTFScoreboardMessage (edict_t *ent, edict_t *killer)
 	gclient_t	*cl;
 	edict_t		*cl_ent;
 	int team;
-	int maxsize = 1000;
+	size_t maxsize = 1000;
 
 	// sort the clients by team and by score within team
 	total[0] = total[1] = 0;
